Limit Magick.NET version for now

+++ tests/ImageSharp.Benchmarks/LoadResizeSave/LoadResizeSaveStressRunner.cs
@@ -248,10 +248,10 @@ public class LoadResizeSaveStressRunner
     public void MagickResize(string input)
     {
         using MagickImage image = new(input);
-        this.LogImageProcessed((int)image.Width, (int)image.Height);
+        this.LogImageProcessed(image.Width, image.Height);
 
         // Resize it to fit a 150x150 square
-        image.Resize((uint)this.ThumbnailSize, (uint)this.ThumbnailSize);
+        image.Resize(this.ThumbnailSize, this.ThumbnailSize);
 
         // Reduce the size of the file
         image.Strip();
